Fix `time` in advisory for squizlabs/php_codesniffer < 2.8.1
The date/time of the issue was in an incorrect format (`2017-01-03` vs `2017-03-01`).

As I was fixing this anyway, I checked the actual time of the commit fixing this which can be seen as the first public announcement of the vulnerability.

So the new time is based on when the fix was originally merged into PHPCS: https://github.com/squizlabs/PHP_CodeSniffer/commit/029305e754dde022c4c8ee48d392c16b61fffe2a