Add a validator ensuring that each branch has a well-formed range
It does not make sense to have multiple upper bounds for a range, as
they would need to match all, meaning that the lowest version of them
would be the useful one. Same is true for lower bounds (except that
the highest version is the meaningful one).
Defining multiple affected ranges is what branches are representing in
the advisory format.